[PT-5] network move validation to core add navigation dependency to core remove navigation dependencies from features
