+++ core/pubspec.yaml
@@ -21,6 +21,8 @@ dependencies:
   hive: ^2.2.3
   hive_flutter: ^1.1.0
   url_launcher: ^6.1.12
+  navigation:
+    path: ../navigation
 
 dev_dependencies:
   flutter_test:
