+++ features/authentication_view/lib/src/ui/auth_screen.dart
@@ -5,7 +5,6 @@ import 'package:core/core.dart';
 import 'package:core_ui/core_ui.dart';
 import 'package:domain/domain.dart';
 import 'package:flutter/material.dart';
-import 'package:navigation/navigation.dart';
 
 class AuthScreen extends StatelessWidget{
   const AuthScreen({super.key});
