+++ features/detailed_dish_view/pubspec.yaml
@@ -11,8 +11,6 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  navigation:
-    path: ../../navigation
   domain:
     path: ../../domain
   core:
