+++ features/main_view/lib/src/ui/main_view_screen.dart
@@ -1,7 +1,6 @@
 import 'package:core/core.dart';
 import 'package:domain/domain.dart';
 import 'package:flutter/material.dart';
-import 'package:navigation/navigation.dart';
 import '../bloc/bloc.dart';
 
 class MainViewScreen extends StatelessWidget {
