+++ features/main_view/pubspec.yaml
@@ -15,8 +15,6 @@ dependencies:
     path: ../../core
   domain:
     path: ../../domain
-  navigation:
-    path: ../../navigation
 
 dev_dependencies:
   flutter_test:
