+++ features/menu_view/lib/src/bloc/menu_view/bloc.dart
@@ -2,7 +2,6 @@ import 'package:core/core.dart';
 import 'package:domain/domain.dart';
 import 'package:flutter/material.dart';
 import 'package:meta/meta.dart';
-import 'package:navigation/navigation.dart';
 
 part 'event.dart';
 part 'state.dart';
