+++ features/shopping_cart_view/pubspec.yaml
@@ -17,8 +17,6 @@ dependencies:
     path: ../../core_ui
   domain:
     path: ../../domain
-  navigation:
-    path: ../../navigation
 
 dev_dependencies:
   flutter_test:
