+++ lib/app/food_app.dart
@@ -2,7 +2,6 @@ import 'package:core/core.dart';
 import 'package:core_ui/core_ui.dart';
 import 'package:domain/domain.dart';
 import 'package:flutter/material.dart';
-import 'package:navigation/navigation.dart';
 import 'package:settings_view/settings_view.dart';
 
 class FoodApp extends StatelessWidget {
