+++ lib/main.dart
@@ -1,7 +1,6 @@
 import 'package:core/core.dart';
 import 'package:data/data.dart';
 import 'package:flutter/material.dart';
-import 'package:navigation/navigation.dart';
 import 'app/food_app.dart';
 
 void main() async{
